Move some domain related functionality to DomainStore
