+++ src/stores/ExploreDataStore.ts
@@ -9,7 +9,7 @@ import { makePersistable } from "mobx-persist-store";
 import { ApeTaxTuple } from "./TaxStore";
 
 /**
- * TODO: Default inputs should be read from the domain configuration file. 
+ * TODO: Default inputs should be read from the domain configuration file.
  */
 const emptyWorkflowConfig = () => {
   return {
@@ -69,17 +69,6 @@ const emptyWorkflowConfig = () => {
   };
 };
 
-interface DomainConfig {
-  ontology_path: string;
-  ontologyPrefixIRI: string;
-  tool_annotations_path: string;
-  strict_tool_annotations: string;
-  use_workflow_input: string;
-  use_all_generated_data: string;
-  tool_seq_repeat: string;
-  constraints_path: string;
-}
-
 export class ExploreDataStore {
   workflowConfig: WorkflowConfig = emptyWorkflowConfig();
   workflowSolutions: WorkflowSolution[] = [];
@@ -116,88 +105,6 @@ export class ExploreDataStore {
     return Object.entries(param).reduce((obj, [key, data]) => {
       return { ...obj, [key]: [data.id] };
     }, {});
-  }
-
-  /**
-   * Reads a domain configuration JSON from a URL and returns it as an object.
-   * @param url The URL of the domain configuration JSON file.
-   * @returns The domain configuration object or undefined in case of an error.
-   */
-  async readDomainConfig(
-    url: string | undefined
-  ): Promise<DomainConfig | null> {
-    if (!url) {
-      console.error("Domain configuration URL is undefined.");
-      return null;
-    }
-    try {
-      const response = await fetch(url, { method: "GET" });
-      if (!response.ok) {
-        throw new Error(
-          `Failed to fetch domain configuration: ${response.status} ${response.statusText}`
-        );
-      }
-      const json = await response.json();
-      if (isDomainConfig(json)) {
-        return json;
-      } else {
-        console.error(
-          "Fetched JSON is not a valid APE domain configuration:",
-          json
-        );
-        return null;
-      }
-    } catch (error) {
-      console.error("Error fetching domain configuration:", error);
-      return null;
-    }
-  }
-
-  /**
-   * Fetches constraints from a given URL and returns them as an array of ConstraintInstance objects.
-   * @param constraintsUrl The URL of the constraints JSON file.
-   * @returns An array of ConstraintInstance objects or undefined in case of an error.
-   */
-  async fetchConstraints(constraintsUrl: string | undefined): Promise<ConstraintInstance[] | undefined> {
-    if (!constraintsUrl) {
-      console.error("Constraints URL is undefined.");
-      return undefined;
-    }
-    try {
-      const response = await fetch(constraintsUrl); // Await the response
-      if (!response.ok) {
-        throw new Error(
-          `Failed to fetch constraints: ${response.status} ${response.statusText}`
-        );
-      }
-      const json = await response.json(); // Await the JSON parsing
-      console.log('Fetched JSON:', json); // Log to inspect the fetched data
-  
-      return json.constraints.map((constraint: any) => {
-        return {
-          id: constraint.constraintid,
-          label: constraint.constraintid,
-          parameters: constraint.parameters.map((param: any) => {
-            return Object.entries(param).reduce((obj, [key, data]) => {
-              if (Array.isArray(data)) {
-                if (data.length === 1 && data[0] !== null) {
-                  // If the data array has a single non-null value, use it
-                  return { ...obj, [key]: data[0] };
-                } else {
-                  // If the array has multiple items or is null, log it
-                  console.warn(`Unexpected data format for ${key}:`, data);
-                  return { ...obj, [key]: null }; // You can handle this case differently if needed
-                }
-              }
-              return { ...obj, [key]: data }; // In case data is not an array
-            }, {});
-          }),
-        };
-      });
-    } catch (error) {
-      console.error("Error fetching constraints:", error);
-      return undefined;
-    }
   }
 
   /**
@@ -267,9 +174,9 @@ export class ExploreDataStore {
       config.domain?.repo_url
     );
 
-    const run_constraints: ConstraintInstance [] = []
-      // The execution is not working with the constraints pulled from the domain config
-      // (await this.fetchConstraints(default_domain_config?.constraints_path)) || [];
+    const run_constraints: ConstraintInstance[] = [];
+    // The execution is not working with the constraints pulled from the domain config
+    // (await this.fetchConstraints(default_domain_config?.constraints_path)) || [];
     config.constraints?.forEach((constraint) => {
       if (constraint.id !== "") {
         run_constraints?.push(constraint);
@@ -383,16 +290,3 @@ export class ExploreDataStore {
 
 const exploreDataStore = new ExploreDataStore();
 export default exploreDataStore;
-function isDomainConfig(obj: any): obj is DomainConfig {
-  return (
-    typeof obj === "object" &&
-    typeof obj.ontology_path === "string" &&
-    typeof obj.ontologyPrefixIRI === "string" &&
-    typeof obj.tool_annotations_path === "string" &&
-    typeof obj.strict_tool_annotations === "string" &&
-    typeof obj.use_workflow_input === "string" &&
-    typeof obj.use_all_generated_data === "string" &&
-    typeof obj.tool_seq_repeat === "string" &&
-    typeof obj.constraints_path === "string"
-  );
-}
